Add specific page for recording calls
Instead of joining the call in the main Talk page now a specific page to
be used only by the recording server was added.

This will make possible to use a specific UI tailored for call
recording, but also to use an internal client of the external signaling
server rather than a normal participant.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>