Use strict dependencies for Debian packages
The Python dependencies are translated into Debian packages by
dh_python3. As described in its man page, by default dh_python3 discards
the version information in the Python dependencies, so those
dependencies that need a strict version have to be explicitly overriden
in a "debian/py3dist-overrides" file.

That file is not included by default in Python source packages, so the
full "debian" directory is explicitly included using a MANIFEST.in file.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>