Add failed status to recordings
Although recordings should not fail in some cases they could (for
example, if the configuration is wrong, the disk is full...). A new
recording status, "failed", has been added to notify that situation to
the Nextcloud server from the recording server.

Typically call participants will not be able to do anything about the
failure, so the clients can just show a notification to the moderators
and hint about asking the system administrators.

In any case, even if the recording status is "failed" further recordings
can be started, which will behave as if the recording was started from
the "none" status.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>