Copy deb package to specific output directory
By default the .deb package is created deep in the build directory, so
for convenience it is now copied to "build/deb".

The output directory is guaranteed to be created when the package is
built by using an "order-only" prerequisite, so it will be created if
needed, but its modification time will not trigger a rebuilt.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>