Include OS version in package name
This will make possible to keep different packages for different
operating system versions in the same repository.

The build directory now also takes into account the operating system
where it was built to be able to share the same directory in a clean
way.

Note that it is expected that "/etc/os-release" is available to find out
the operating system version.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>