Add helper script to build the packages
The script creates one container for each supported distribution,
installs the needed dependencies and then builds the packages.

The recording server directory is mounted in the containers so the
packages can be built in the "build" subdirectory of the actual
recording server directory. In the containers the building is done as
the owner of the recording server directory, so the built files are also
owned by that user.

For simplicity, and given that the containers are not expected to be
used for anything else, the volume with the recording server is mounted
as a direct child of the root directory (forgive me FSH for I have
sinned).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>