Abort starting the helpers when the recording is stopped
The recording is stopped in a different thread than it is started.
Stopping the recording stops the helpers, like the display, but if they
are stopped at the same time that they are being started there could be
a race condition in which the helper is tried to be stopped in one
thread and, immediately after that, the helper is started in the other
thread. This is now handled by explicitly checking if the recording was
stopped after starting each helper and aborting the start if it was.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>