Accept insecure certificates in the browser if "skipverify" is enabled
"skipverify" can be set in the configuration to skip certificate
validation of the backend endpoints. However, this was honoured only for
requests done by the Python code; if the Nextcloud server was using an
invalid certificate, like a self-signed one, the browser launched by the
recording server was not be able to load the page that acts as its entry
point in the Nextcloud server, and therefore the recording failed.

Note that this was an issue only in Chromium because the Selenium Python
bindings accept insecure certificates by default in Firefox.
Nevertheless now "skipverify" will be honoured too when disabled, so
Firefox would also fail if "skipverify = false" and Nextcloud uses an
invalid certificate (although it is likely that Firefox would not be
even launched in that case, as the failed request sent to the Nextcloud
server by the Python code would make the recording fail early).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>