Add dependency and executable for the benchmark tool
The benchmark tool has a specific dependency (psutil) that is not
required for the rest of the recording server. However, for simplicity,
rather than splitting the benchmark in its own Python project and
package its dependency and helper script is now included in the main
project and package (the code was already there).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>