Show frames dropped by ffplay in benchmark summary
The dropped frames are shown by default by ffplay, but ffplay output is
only shown when the benchmark is run with "--verbose-extra". However,
knowing whether some frames were dropped by the player is useful, so
now the final count of dropped frames is shown in the summary
independently of the verbosity level.

The dropped frames are parsed from ffplay output based on its current
format, although it has not changed for years:
https://git.ffmpeg.org/gitweb/ffmpeg.git/blob/4acd08be6c4f39736179a3d90fd56b508e42ff6d:/fftools/ffplay.c#l1705

The log of the ffplay process was consumed in its own thread by
"processLog". Therefore it was not possible to just add another thread
to track the dropped frames, as reading removes the contents and thus
each thread would only get part of the output. Unfortunately I did not
find an easy way to duplicate the TextIOWrapper to be read from two
different threads (stdout would still need to be read by one thread
which then would duplicate the contents to two different TextIOWrappers
each one in turn read by a different thread). For simplicity, for now a
custom method is used to handle the ffplay output and both log the
output and keep track of the dropped frames.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>