Add argument to overwrite the benchmark output file
By default the benchmark tool refuses to overwrite the output file it if
exists to prevent data loss. However, in some cases (for example, when
trying whether some ffmpeg options are valid or not) it is useful to
automatically overwrite the existing file instead of having to manually
delete it before trying another option.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>