Rename package to "nextcloud-talk-recording"
By default stdeb prefixes the package name with "python3-". However,
even if other Python modules could import the Python code of the
recording server it is not designed as a library, it is an executable
that happens to be written in Python, so to speak.

Moreover, besides the Python code itself the package also includes a
configuration file and a systemd service, which do not belong to
"python3-" packages (there are also helper executable scripts, although
those are allowed in "python3-" packages).

Due to all that the package "python3-nextcloud-talk-recording" is now
renamed to "nextcloud-talk-recording".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>