Enforce installing from local directory instead of from PyPi
When a "plain" name is given to "pip install" it looks for matching
projects on PyPi and local directories and then installs the best one in
terms of version number.

To ensure that the name is seen as a local directory and directly
installed from it rather than also checking PyPi the name must provide
any hint of being a local directory. Therefore, if the name is given as
"nextcloud-talk-recording/" or "./nextcloud-talk-recording" it should be
enough. However, to err on the safe side the longer specification of
'"file://$(pwd)/nextcloud-talk-recording"' was used instead (the double
quotes also ensure that if there are spaces in the parent directories
they will be properly processed).

For consistency "file://" was also added to other uses of "pip install",
even if it was not strictly necessary.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>