Add configuration for trusted proxies
Note that, unlike the equivalent configuration option in the signaling
server, in the recording server there is no default value. Any trusted
proxy, even on the loopback interface, needs to be explicitly
configured; as the trusted proxies must set the "X-Forwarded-For" header
no defaults are provided, as a reverse proxy in front may pass rather
than set the header if not properly setup.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>