Add middleware to adjust the remote address based on the trusted proxies
The WSGI middleware adjusts "REMOTE_ADDRESS" based on the remote
address, the "X-Forwarded-For" header and the list of trusted proxies,
so now, if everything is properly configured, the "real" client IP is
seen by the Flask application.

pytest version had to be increased to at least 6.0.1, as in previous
versions pylint incorrectly complained that "pytest.mark.parametrize is
not callable".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>