Pin setuptools < 69.3 to keep the old Python source package name
Starting with setuptools 69.3 support for PEP 625 was implemented, so
the name of the source distribution is now "nextcloud_talk_recording"
instead of "nextcloud-talk-recording". The Makefile expects the later,
so the setuptools used to generate the Python source package needs to be
explicitly pinned to < 69.3. Otherwise, as no build system was specified
in the pyproject.toml file, "build" would use "setuptools >= 40.8.0" by
default, which would install the latest version available, and therefore
>= 69.3. setuptools 61 is used as the lowest version, as it is required
for proper autodiscovery of the project layout (so the fact that it had
worked until now without an explicit version was a happy coincidence due
to installing the latest available >= 40.8.0).

Strictly speaking the requirement of using setuptools < 69.3 should not
be in the pyproject.toml file, as it is not an actual requirement to
build the source distribution of the project in general but only when
used with the Makefile to build the distribution packages. But for
simplicity it will set like that, at least for now.

Note that the setuptools installed system wide in "build.sh", which is
used by "stdeb" to build the Debian binary package, does not need to be
pinned, as "stdeb" does not expect the source package to have a specific
name.

However, in the specific case of Ubuntu 22.04 it must be, as due to a
bug in python3-build the virtual environment used to build the source
distribution is not actually isolated but gets its build dependencies
from those installed system wide.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>