Remove unneeded visit to main Nextcloud server page
This is just an unneeded leftover from the initial experiments with
recordings. Moreover, it could prevent recordings from starting, for
example if SAML is enabled in the Nextcloud server and the identity
provider is not accesible from the recording server, as an exception
would be thrown in that case when creating the Participant object.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>