feat: Add support for specifying Selenium driver and browser executable
Since version 4.6 Selenium uses the Selenium Manager to discover the
Selenium driver in $PATH, downloading it as a fallback. However, the
Selenium Manager does not work in certain architectures, like Linux on
arm64/aarch64, so in those architectures when Selenium started it was
not able to find the driver and the recording failed. Now it is possible
to explicitly set the path to the driver, which ensures that Selenium
can find it and thus start, even if Selenium Manager is not available.

Besides that, even if Selenium Manager is available, being able to set
the path to the driver is also needed in cases in which the driver is
not in a standard location.

Similarly, the Selenium Manager is also used to discover the browser
executable in $PATH, downloading it as a fallback (since version 4.11
for "chrome" and 4.12 for "firefox"). The Selenium driver also looks for
the browser executable, so even if Selenium Manager is not available it
might still be automatically found. Nevertheless, like with the driver,
in some cases it is also necessary to be able to set a specific path.

Signed-off-by: cb0s <cedric.boes@online.de>