fix: Set a unique name for each Pylint check
The lint action is used several times in the same workflow due to being
run using a matrix of Python versions. Due to that the check created by
it needs to have a unique name for each run, or otherwise once a new run
finished the previous check was overwritten.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>