Add metric for total duration of all recordings
It would be possible to get the exact duration of each recording by
parsing the output of ffmpeg. However, that would require increasing the
log level of ffmpeg to "info", which would either cause the default logs
of the recording server to be too verbose or require some special
handling to filter them by default and yet allow admins to print them if
desired (as it is currently possible to print them by increasing the log
level of ffmpeg in the configuration). Therefore, for simplicity, the
duration is calculated from the start and stop of the service, even if
that might introduce a difference of a few seconds with the actual
duration of the recording file.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>