Fix building with stdeb 0.11.0
stdeb 0.11.0 raised the minimum debhelper version to 12. However, it
still uses "--with systemd", which is only supported in debhelper <= 10,
so building the nextcloud-talk-recording package fails.

Besides that, it also changed the build system from python_distutils to
pybuild, which causes the build of Selenium to fail.

Due to all that, stdeb is pinned to < 0.11.0.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>