Fix Selenium manager permissions in selenium-4.14.0 package
The Python package for Selenium 4.14.0 provides a pre-built
selenium-manager, but in the source package it does not have execute
permissions (it does on the binary, though). As the file is just copied
as is from the source Python package to the .deb package it is necessary
to explicitly set the execute permissions.

Note that this implies that an executable for a specific architecture is
included in a .deb package for any architecture. While that is indeed
incorrect for now it will be kept as is, as it is not possible to build
selenium-manager for Linux on other architectures other than amd64, and
using Selenium manager can be skipped by explicitly providing a path to
the driver in the recording server configuration.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>