Bump requests version to 2.32.4
Ideally the version should be bumped to 2.32.5, as that version fixes a
regression introduced in 2.32.0 with the caching of SSLContext (by
reverting that feature). Unfortunately requests 2.32.5 requires Python
3.9, which is not available in Ubuntu 20.04, so the latest 2.32.X
compatible version is used instead.

On the bright side the bugs introduced by the caching do not seem to be
problematic in the case of the recording server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>