Build charset-normalizer in Ubuntu 20.04
Since requests 2.26 chardet was replaced with charset-normalizer as an
explicit dependency in Python 3 (although even in latest version it
seems to still work if charset-normalizer is not available but chardet
is). However, in 2.32.0 it became an optional dependency and instead of
failing to start if the libraries are not there it now shows a warning.

Nevertheless, although the character detection is probably not needed in
the recording server, charset-normalizer is now packaged for Ubuntu
20.04 to avoid that warning (and also just in case the character
detection turns out to be actually needed).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>